refactor(vault): split entry feature cards and address review feedback
- Split DisconnectedFeatureCards into one component per file
- Use useId() for the partial-liquidation mask to avoid duplicate DOM ids
- Memoize aprStats and featureCards in DisconnectedOverview
- Correct the docstring: the rates card shows live APRs statically, only the last two cards expand
