+++ services/vault/src/components/simple/DisconnectedFeatureCards/CompetitiveRatesIcon.tsx
@@ -1,0 +1,27 @@
+import { ICON_SIZE } from "./iconSize";
+
+export function CompetitiveRatesIcon() {
+  return (
+    <svg
+      width={ICON_SIZE}
+      height={ICON_SIZE}
+      viewBox="0 0 32 32"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      aria-hidden="true"
+    >
+      <path
+        d="M12.8598 20C14.369 20 15.5904 18.7791 15.5904 17.2706C15.5904 15.7621 14.369 14.5412 12.8598 14.5412C11.3507 14.5412 10.1293 15.7621 10.1293 17.2706C10.1293 18.7791 11.3507 20 12.8598 20Z"
+        fill="currentColor"
+      />
+      <path
+        d="M19.9966 20C21.5057 20 22.7272 18.7791 22.7272 17.2706C22.7272 15.7621 21.5057 14.5412 19.9966 14.5412C18.4874 14.5412 17.266 15.7621 17.266 17.2706C17.266 18.7791 18.4874 20 19.9966 20Z"
+        fill="currentColor"
+      />
+      <path
+        d="M16.4227 6C9.00791 6 2.9968 12.1235 3 19.6725H6.43081C6.43081 14.0124 10.8688 9.42612 16.4259 9.42612C21.983 9.42612 26.421 14.0156 26.421 19.6725H29.8518C29.8518 12.1203 23.8407 6 16.4291 6H16.4227Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
