+++ services/vault/src/components/simple/DisconnectedFeatureCards/FastAccessIcon.tsx
@@ -1,0 +1,46 @@
+import { ICON_SIZE } from "./iconSize";
+
+export function FastAccessIcon() {
+  return (
+    <svg
+      width={ICON_SIZE}
+      height={ICON_SIZE}
+      viewBox="0 0 32 32"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      aria-hidden="true"
+    >
+      <line
+        x1="3.665"
+        y1="11.335"
+        x2="9.335"
+        y2="11.335"
+        stroke="currentColor"
+        strokeWidth="1.33"
+        strokeLinecap="round"
+      />
+      <line
+        x1="0.665"
+        y1="15.335"
+        x2="8.335"
+        y2="15.335"
+        stroke="currentColor"
+        strokeWidth="1.33"
+        strokeLinecap="round"
+      />
+      <line
+        x1="3.665"
+        y1="19.335"
+        x2="9.335"
+        y2="19.335"
+        stroke="currentColor"
+        strokeWidth="1.33"
+        strokeLinecap="round"
+      />
+      <path
+        d="M20.8765 25.3594C26.2208 25.3594 30.5562 21.024 30.5562 15.6797C30.5562 10.3354 26.2208 6 20.8765 6C15.5322 6 11.1968 10.3354 11.1968 15.6797C11.1968 21.024 15.5322 25.3594 20.8765 25.3594ZM20.8765 23.6512C16.4712 23.6512 12.905 20.085 12.905 15.6797C12.905 11.2744 16.4712 7.70818 20.8765 7.70818C25.2818 7.70818 28.838 11.2744 28.838 15.6797C28.838 20.085 25.2818 23.6512 20.8765 23.6512ZM25.7613 16.7485C26.1508 16.7485 26.4505 16.4489 26.4505 16.0593C26.4505 15.6797 26.1508 15.38 25.7613 15.38H21.5657V9.74601C21.5657 9.36641 21.2661 9.06673 20.8865 9.06673C20.4969 9.06673 20.1872 9.36641 20.1872 9.74601V16.0593C20.1872 16.4489 20.4969 16.7485 20.8865 16.7485H25.7613Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
