+++ services/vault/src/components/simple/DisconnectedFeatureCards/PartialLiquidationIcon.tsx
@@ -1,0 +1,33 @@
+import { useId } from "react";
+
+import { ICON_SIZE } from "./iconSize";
+
+export function PartialLiquidationIcon() {
+  const maskId = useId();
+
+  return (
+    <svg
+      width={ICON_SIZE}
+      height={ICON_SIZE}
+      viewBox="0 0 32 32"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      aria-hidden="true"
+    >
+      <mask id={maskId} fill="white">
+        <path d="M27.1781 16.8193C27.1781 18.908 26.5587 20.9497 25.3983 22.6864C24.2379 24.423 22.5886 25.7766 20.6589 26.5759C18.7292 27.3752 16.6059 27.5843 14.5573 27.1768C12.5088 26.7694 10.6271 25.7636 9.15022 24.2867C7.67331 22.8098 6.66752 20.9281 6.26005 18.8795C5.85257 16.831 6.0617 14.7077 6.861 12.778C7.66029 10.8483 9.01385 9.19899 10.7505 8.03859C12.4872 6.87819 14.5289 6.25883 16.6176 6.25883V16.8193H27.1781Z" />
+      </mask>
+      <path
+        d="M27.1781 16.8193C27.1781 18.908 26.5587 20.9497 25.3983 22.6864C24.2379 24.423 22.5886 25.7766 20.6589 26.5759C18.7292 27.3752 16.6059 27.5843 14.5573 27.1768C12.5088 26.7694 10.6271 25.7636 9.15022 24.2867C7.67331 22.8098 6.66752 20.9281 6.26005 18.8795C5.85257 16.831 6.0617 14.7077 6.861 12.778C7.66029 10.8483 9.01385 9.19899 10.7505 8.03859C12.4872 6.87819 14.5289 6.25883 16.6176 6.25883V16.8193H27.1781Z"
+        stroke="currentColor"
+        strokeWidth="2.66"
+        mask={`url(#${maskId})`}
+      />
+      <path
+        d="M27.0459 15.0603C26.9896 14.0414 26.7625 13.0376 26.3711 12.0926C25.9158 10.9934 25.2475 9.99535 24.4062 9.15408C23.565 8.31282 22.5669 7.64457 21.4678 7.18924C20.5228 6.79781 19.5189 6.57173 18.5 6.51541V15.0603H27.0459Z"
+        fill="currentColor"
+        stroke="currentColor"
+      />
+    </svg>
+  );
+}
