+++ services/vault/src/components/simple/DisconnectedFeatureCards/SelfCustodialIcon.tsx
@@ -1,0 +1,19 @@
+import { ICON_SIZE } from "./iconSize";
+
+export function SelfCustodialIcon() {
+  return (
+    <svg
+      width={ICON_SIZE}
+      height={ICON_SIZE}
+      viewBox="0 0 32 32"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      aria-hidden="true"
+    >
+      <path
+        d="M23.9582 13.4424C24.2911 11.2358 22.5967 10.0496 20.2799 9.25825L21.0314 6.26913L19.1964 5.81569L18.4648 8.72603C17.9824 8.60684 17.4869 8.49439 16.9946 8.38297L17.7315 5.45345L15.8976 5L15.1455 7.98808C14.7462 7.89791 14.3542 7.80878 13.9737 7.71498L13.9758 7.70565L11.4452 7.07912L10.9571 9.02246C10.9571 9.02246 12.3185 9.33184 12.2898 9.35101C13.033 9.53498 13.1673 10.0226 13.1448 10.4092L12.2888 13.8145C12.34 13.8274 12.4063 13.8461 12.4795 13.8751C12.4184 13.8601 12.353 13.8435 12.2856 13.8274L11.0856 18.5977C10.9947 18.8216 10.7642 19.1574 10.2447 19.0299C10.263 19.0563 8.91096 18.6998 8.91096 18.6998L8 20.7825L10.3879 21.3728C10.8322 21.4832 11.2675 21.5987 11.6961 21.7076L10.9367 24.7309L12.7696 25.1843L13.5217 22.1931C14.0223 22.3279 14.5084 22.4522 14.984 22.5694L14.2345 25.5466L16.0695 26L16.8289 22.9824C19.9579 23.5695 22.3108 23.3327 23.3012 20.5265C24.0993 18.2671 23.2615 16.9637 21.6152 16.1138C22.8141 15.8397 23.7172 15.0577 23.9582 13.4424ZM19.7656 19.2719C19.1985 21.5314 15.3619 20.3099 14.118 20.0036L15.1256 15.9983C16.3695 16.3061 20.3583 16.9155 19.7656 19.2719ZM20.3332 13.4097C19.8158 15.465 16.6225 14.4208 15.5866 14.1648L16.5002 10.532C17.536 10.788 20.872 11.2659 20.3332 13.4097Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
