+++ services/vault/src/components/simple/DisconnectedFeatureCards/iconSize.ts
@@ -1,0 +1,1 @@
+export const ICON_SIZE = 32;
